pyproject.toml
[tool.poetry]
name = "ecommerce"
version = "0.1.0"
description = ""
authors = ["eedriz99 <akinsolaidris1999@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = {extras = ["argon2"], version = "^5.0.2"}
django-cors-headers = "4.3.1"
djangorestframework = "3.14.0"
pillow = "^10.2.0"
drf-social-oauth2 = "2.1.3"
djangorestframework-simplejwt = "^5.3.1"
whitenoise = {extras = ["brotli"], version = "^6.6.0"}
psycopg2-binary = "^2.9.9"
dj-database-url = "^2.1.0"
gunicorn = "^21.2.0"
uvicorn = "^0.27.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
